+++ source/kernel/timer.h
@@ -30,22 +30,22 @@
 class Timer {
  public:
   virtual bool Setup() = 0;
-  virtual volatile uint32_t ReadMainCnt() = 0;
+  virtual volatile uint64_t ReadMainCnt() = 0;
   // us秒後のカウントを取得する
-  volatile uint32_t GetCntAfterPeriod(volatile uint32_t cur, uint32_t us) {
-    return cur + (us * 1000) / _cnt_clk_period;
+  volatile uint64_t GetCntAfterPeriod(volatile uint64_t cur, int us) {
+    return cur + (static_cast<int64_t>(us) * 1000) / _cnt_clk_period;
   }
-  volatile bool IsTimePassed(volatile uint32_t time) {
-    volatile uint32_t cur = ReadMainCnt();
+  volatile bool IsTimePassed(volatile uint64_t time) {
+    volatile uint64_t cur = ReadMainCnt();
     return cur >= time;
   }
   // ビジーループタイムアウト
   // 最適化回避のためにできるだけこの関数を使うべき
-  void BusyUwait(uint32_t us) {
-    volatile uint32_t cur = ReadMainCnt();
-    volatile uint32_t end = GetCntAfterPeriod(cur, us);
+  void BusyUwait(int us) {
+    volatile uint64_t cur = ReadMainCnt();
+    volatile uint64_t end = GetCntAfterPeriod(cur, us);
     if (cur > end) {
-      for (volatile uint32_t tmp = ReadMainCnt(); cur <= tmp; tmp = ReadMainCnt()) {
+      for (volatile uint64_t tmp = ReadMainCnt(); cur <= tmp; tmp = ReadMainCnt()) {
       }
     }
     while(true) {
@@ -55,15 +55,15 @@ class Timer {
       }
     }
   }
-  uint32_t GetCntClkPeriod() {
+  uint64_t GetCntClkPeriod() {
     return _cnt_clk_period;
   }
-  int64_t GetUsecFromCnt(int32_t cnt) {
-    return (static_cast<int64_t>(cnt) * _cnt_clk_period) / 1000;
+  uint64_t GetUsecFromCnt(uint64_t cnt) {
+    return (cnt * _cnt_clk_period) / 1000;
   }
  protected:
   // １カウントが何ナノ秒か
-  uint32_t _cnt_clk_period = 1;
+  uint64_t _cnt_clk_period = 1;
 };
 
 #endif /* __RAPH_KERNEL_TIMER_H__ */
